Enable out of order tracking of requests
With the added support of server-speaks-by-itself it is possible that
server response reaches the tracker before the client request. For
example this could happen when there is a scheduling delay while the
client is waiting for the message to be sent to th tracker.

Enforce client->server ordering by sequencing the messages in the
tracker again. Server speaks first situations are not supported and
will cause the tracker to stop.